fix: make MAX_RESPONSE_CHUNK_SIZE importable; fix Cookie.__init__ max_age param type
- scribe.pyx: cdef int MAX_RESPONSE_CHUNK_SIZE is not exported as a Python name
  from .so files; expose as Python-level constant and use private C alias internally
- cookies.pyx: Cookie.__init__ param was still typed as C int, causing OverflowError
  before assignment to the long long field; changed param to long long

+++ blacksheep/cookies.pyx
@@ -40,7 +40,7 @@ cdef class Cookie:
         str path=None,
         bint http_only=0,
         bint secure=0,
-        int max_age=-1,
+        long long max_age=-1,
         CookieSameSiteMode same_site=CookieSameSiteMode.UNDEFINED
     ):
         self.name = name
